+++ frontend/src/pages/Graphing/Dashboard.tsx
@@ -57,7 +57,7 @@ import { useRetentionPresets } from '@/components/Search/SearchForm/hooks'
 import { loadFunnelStep } from '@pages/Graphing/util'
 import { GraphContextProvider } from './context/GraphContext'
 import { useGraphData } from '@pages/Graphing/hooks/useGraphData'
-import { useExportGraph } from '@pages/Graphing/hooks/useExportGraph'
+import { exportGraph } from '@pages/Graphing/hooks/exportGraph'
 
 export const HeaderDivider = () => <Box cssClass={style.headerDivider} />
 
@@ -170,7 +170,6 @@ export const Dashboard = () => {
 	const tempId = useId()
 
 	const graphContext = useGraphData()
-	const { exportGraph } = useExportGraph()
 
 	const noGraphs = graphs?.length === 0
 
@@ -549,6 +548,10 @@ export const Dashboard = () => {
 																	g.title,
 																	g.functionType,
 																	g.metric,
+																	graphContext
+																		.graphData[
+																		g.id
+																	],
 																)
 															}
 														>
